+++ src/org/rascalmpl/value/io/binary/util/FieldKind.java
@@ -14,7 +14,7 @@ package org.rascalmpl.value.io.binary.util;
 
 public class FieldKind {
     public static final int PREVIOUS_STR = 1;
-    public static final int LONG = 2;
+    public static final int INT = 2;
     public static final int STRING = 3;
     public static final int BYTES = 4;
     public static final int REPEATED = 5;
